Balança/requirements.txt
# Para o Dashboard e manipulação de dados
dash==2.17.1
dash-bootstrap-components==1.6.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
openpyxl==3.1.2

# Para comunicação com a porta serial
pyserial==3.5

# Como usar
# pip install -r requirements.txt